fix: search's configuration about buttonAriaLabel doesn't work
